Enable get_number_open_files on more platforms, and use that to track down the refcycle in FileObjectThread.
