+++ dev-requirements.txt
@@ -7,6 +7,7 @@ coverage>=4.0
 coveralls>=1.0
 cffi
 futures
+psutil
 # For viewing README.rst (restview --long-description),
 # CONTRIBUTING.rst, etc.
 # https://github.com/mgedmin/restview
